requirements.txt
Django==1.8.2
pinax-theme-bootstrap==5.7.4
django-user-accounts==1.1
metron==1.3.5
pinax-eventlog==1.0.0
pinax-notifications==2.1.0
django-jsonfield==0.9.13
psycopg2==2.6
gondor==1.2.5
gunicorn==19.3.0
opbeat==2.1.1
-e git+https://github.com/ntoll/pinax-teams.git@bd25f2442f9f87763749c239ae3d2bd7bf2589e0#egg=pinax_teams-compat-be-gone
kaleo
-e git+https://github.com/pinax/pinax-documents.git@d5a86ee7a44e2a93083db66536d860ce9d276acb#egg=pinax_documents
